+++ vs/src/main.rs
@@ -38,6 +38,7 @@ mod router;
 mod signal_worker;
 mod topology_mgr;
 mod trusted_services;
+mod visa_bootstrap;
 mod visa_mgr;
 mod visa_policy;
 mod visa_reconciler;
